+++ CELIKOOR_PINKMAN/FormDaftarGenre.cs
@@ -49,34 +49,42 @@ namespace CELIKOOR_PINKMAN
 
         private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
+            if (e.ColumnIndex == dataGridView1.Columns["btnUbahGrid"].Index && e.RowIndex >= 0)
+            {
+
+            }
+            if (e.ColumnIndex == dataGridView1.Columns["btnDeleteGrid"].Index && e.RowIndex >= 0)
+            {
+                if (dataGridView1.SelectedRows.Count > 0)
+                {
+                    string kodeHapus = dataGridView1.CurrentRow.Cells["ID"].Value.ToString();
+                    string namaHapus = dataGridView1.CurrentRow.Cells["Nama"].Value.ToString();
+
+                    DialogResult hasil = MessageBox.Show(this, "Anda yakin akan menghapus?" + kodeHapus + "-" +
+                        namaHapus + "?", "HAPUS", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+                    if (hasil == DialogResult.Yes)
+                    {
+                        Aktor f = Aktor.SelectDataSingle(kodeHapus);
+                        Boolean hapus = Aktor.DeleteData(f);
+
+                        if (hapus)
+                        {
+                            MessageBox.Show("Penghapusan data berhasil");
+
+                        }
+                        else
+                        {
+                            MessageBox.Show("Penghapusan data gagal");
+                        }
+                    }
+                }
+            }
         }
 
         private void buttonDelete_Click(object sender, EventArgs e)
         {
-            if (dataGridView1.SelectedRows.Count > 0)
-            {
-                string kodeHapus = dataGridView1.CurrentRow.Cells["ID"].Value.ToString();
-                string namaHapus = dataGridView1.CurrentRow.Cells["Nama"].Value.ToString();
-
-                DialogResult hasil = MessageBox.Show(this, "Anda yakin akan menghapus?" + kodeHapus + "-" +
-                    namaHapus + "?", "HAPUS", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-
-                if (hasil == DialogResult.Yes)
-                {
-                    Aktor f = Aktor.SelectDataSingle(kodeHapus);
-                    Boolean hapus = Aktor.DeleteData(f);
-
-                    if (hapus)
-                    {
-                        MessageBox.Show("Penghapusan data berhasil");
-
-                    }
-                    else
-                    {
-                        MessageBox.Show("Penghapusan data gagal");
-                    }
-                }
-            }
+            
         }
     }
 }
